requirements.txt
# Core dependencies - Fixed versions for compatibility
langchain>=0.1.0
langchain-google-genai>=0.0.6
langchain-huggingface>=0.3.1
chromadb>=0.4.22
sentence-transformers>=2.2.2

# Pipecat and streaming - Updated to latest version
pipecat-ai==0.0.79
fastapi>=0.104.1
websockets>=12.0
uvicorn>=0.24.0

# Audio processing
soundfile>=0.12.1
numpy>=1.24.0
pydub>=0.25.1
SpeechRecognition>=3.10.0
gTTS>=2.3.2

# Web framework
streamlit>=1.28.0
streamlit-audiorecorder>=0.0.5

# Async and utilities
asyncio-mqtt>=0.11.0
aiofiles>=23.0.0
python-dotenv>=1.0.0
pydantic-settings>=2.0.0

# Google services
google-cloud-speech
google-cloud-texttospeech

# HTTP requests
requests>=2.31.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0